requirements.txt
pyfiglet
termcolor
PyYAML==5.3.1
requests==2.25.0